+++ astropy/coordinates/builtin_frames/__init__.py
@@ -50,32 +50,42 @@ from . import cirs_observed_transforms
 from . import intermediate_rotation_transforms
 from . import ecliptic_transforms
 
-# we define an __all__ because otherwise the transformation modules get included
+from ..baseframe import frame_transform_graph
+
+# we define an __all__ because otherwise the transformation modules
+# get included
 __all__ = ['ICRS', 'FK5', 'FK4', 'FK4NoETerms', 'Galactic', 'Galactocentric',
            'Supergalactic', 'AltAz', 'GCRS', 'CIRS', 'ITRS', 'HCRS',
            'PrecessedGeocentric', 'GeocentricTrueEcliptic',
            'BarycentricTrueEcliptic', 'HeliocentricTrueEcliptic',
            'SkyOffsetFrame', 'GalacticLSR', 'LSR',
-           'BaseEclipticFrame', 'BaseRADecFrame']
+           'BaseEclipticFrame', 'BaseRADecFrame', 'make_transform_graph_docs']
 
 
-def _make_transform_graph_docs():
+def make_transform_graph_docs(transform_graph):
     """
-    Generates a string for use with the coordinate package's docstring
-    to show the available transforms and coordinate systems
+    Generates a string that can be used in other docstrings to include a
+    transformation graph, showing the available transforms and
+    coordinate systems.
+
+    Parameters
+    ----------
+    transform_graph : `~.coordinates.TransformGraph`
+
+    Returns
+    -------
+    docstring : str
+        A string that can be added to the end of a docstring to show the
+        transform graph.
     """
-    import inspect
     from textwrap import dedent
-    from ..baseframe import BaseCoordinateFrame, frame_transform_graph
-
-    isclass = inspect.isclass
-    coosys = [item for item in globals().values()
-              if isclass(item) and issubclass(item, BaseCoordinateFrame)]
+    coosys = [transform_graph.lookup_name(item) for
+              item in transform_graph.get_names()]
 
     # currently, all of the priorities are set to 1, so we don't need to show
     #   then in the transform graph.
-    graphstr = frame_transform_graph.to_dot_graph(addnodes=coosys,
-                                                  priorities=False)
+    graphstr = transform_graph.to_dot_graph(addnodes=coosys,
+                                            priorities=False)
 
     docstr = """
     The diagram below shows all of the coordinate systems built into the
@@ -123,4 +133,4 @@ def _make_transform_graph_docs():
     return docstr
 
 
-_transform_graph_docs = _make_transform_graph_docs()
+_transform_graph_docs = make_transform_graph_docs(frame_transform_graph)